task2/requirements.txt
streamlit
tweepy
pandas
matplotlib
vaderSentiment
python-dotenv